Lock the `credit-refill` from being deleted

+++ src/api/models/Admin/Menu/index.ts
@@ -1,15 +1,15 @@
 // Import our proper types
 import { Request, Response } from "express";
 // Import the promotion
+import mongoose from "mongoose";
+import settings from "../../../../config/settings.json";
 import Menu from "../../../../database/models/Products";
 import Reviews from "../../../../database/models/Reviews";
 import { ErrorFormat, iwe_strings } from "../../../strings";
 import { get_authorization_user } from "../../../utility/Authentication";
+import { what_is, wis_array, wis_string } from "../../../utility/What_Is";
 import what from "../../../utility/Whats";
-import { what_is, wis_array } from "../../../utility/What_Is";
 import { delete_object } from "../../../utility/batchRequest";
-import settings from "../../../../config/settings.json";
-import mongoose from "mongoose";
 
 // Create a new product
 async function menu_create(req: Request, res: Response) {
@@ -56,7 +56,7 @@ async function menu_create(req: Request, res: Response) {
     in_stock,
     description,
     category,
-    keywords,
+    keywords
   );
 
   if (testFailed) return;
@@ -72,7 +72,7 @@ async function menu_create(req: Request, res: Response) {
   // There are some reserved slugs that we can't use
   if (
     settings.products["disallowed-product-names"].includes(
-      slug.trim().toLowerCase(),
+      slug.trim().toLowerCase()
     )
   ) {
     return res.status(400).json(ErrorFormat(iwe_strings.Product.ERESERVEDSLUG));
@@ -105,6 +105,12 @@ async function menu_create(req: Request, res: Response) {
 
 // delete menu
 async function menu_delete(req: Request, res: Response) {
+  // Do not allow the "credit-refill" to be deleted
+  const product_slug = wis_string(req);
+  if (product_slug == "credit-refill") {
+    return res.status(401).json(ErrorFormat(iwe_strings.Product.ELOCKED));
+  }
+
   await delete_object(
     req,
     res,
@@ -114,7 +120,7 @@ async function menu_delete(req: Request, res: Response) {
     iwe_strings.Product.ENOTFOUND,
     true,
     Reviews,
-    "product",
+    "product"
   );
 }
 
@@ -165,7 +171,7 @@ async function menu_modify(req: Request, res: Response) {
     description,
     category,
     keywords,
-    old_slug,
+    old_slug
   );
 
   if (testFailed) return;
@@ -179,7 +185,7 @@ async function menu_modify(req: Request, res: Response) {
   // There are some reserved slugs that we can't use
   if (
     settings.products["disallowed-product-names"].includes(
-      slug.trim().toLowerCase(),
+      slug.trim().toLowerCase()
     )
   ) {
     return res.status(400).json(ErrorFormat(iwe_strings.Product.ERESERVEDSLUG));
@@ -214,7 +220,7 @@ async function menu_modify(req: Request, res: Response) {
   await menu.save();
 
   return res.json(
-    what_is(what.private.menu, [iwe_strings.Order.IPMODIFY, menu]),
+    what_is(what.private.menu, [iwe_strings.Order.IPMODIFY, menu])
   );
 }
 
@@ -228,7 +234,7 @@ function check_values(
   description: string,
   category: string,
   keywords: string[],
-  old_slug?: string,
+  old_slug?: string
 ) {
   if (
     (old_slug && typeof old_slug != "string") ||
@@ -254,3 +260,4 @@ function check_values(
 }
 
 export { menu_create, menu_delete, menu_modify };
+
